Move and rename languageSupport to language-support
